libft header

+++ src/ft_strlcat.c
@@ -6,11 +6,15 @@
 /*   By: ydidenko <ydidenko@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2023/10/07 14:21:26 by ydidenko      #+#    #+#                 */
-/*   Updated: 2023/10/10 17:46:17 by ydidenko      ########   odam.nl         */
+/*   Updated: 2023/10/10 18:14:19 by ydidenko      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
-unsigned int	ft_strlcat(char *dest, char *src, unsigned int size)
+#include "stddef.h"
+
+
+
+size_t	ft_strlcat(char *dest, char *src, size_t size)
 {
 	int	count;
 
@@ -27,5 +31,6 @@ unsigned int	ft_strlcat(char *dest, char *src, unsigned int size)
 		}
 		count++;
 	}
+
 	return (count);
 }
