+++ src/ft_strlen.c
@@ -6,11 +6,13 @@
 /*   By: ydidenko <ydidenko@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2023/10/06 13:32:13 by ydidenko      #+#    #+#                 */
-/*   Updated: 2023/10/10 17:46:27 by ydidenko      ########   odam.nl         */
+/*   Updated: 2023/10/10 18:08:51 by ydidenko      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(char *str)
+#include "stddef.h"
+
+size_t	ft_strlen(char *str)
 {
 	int	i;
